Refactor Login & Sign UP Screen For Bloc Consumer To Mange Send Notification
